nene-backend/requirements.txt

flask
flask-cors
pymongo
python-dotenv
werkzeug
